# docs/swagger.yaml
definitions:
  controller.response:
    properties:
      message:
        type: string
    type: object
...
  models.Admins:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isSuperuser:
        type: boolean
      lastVisit:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  models.Banner:
    properties:
      created:
        $ref: '#/definitions/models.Admins'
      createdAt:
        type: string
      descriptionEn:
        type: string
      descriptionRu:
        type: string
      descriptionUz:
        type: string
      id:
        type: integer
      image:
        type: string
      nameEn:
        type: string
      nameRu:
        type: string
      nameUz:
        type: string
      position:
        type: integer
      updated:
        $ref: '#/definitions/models.Admins'
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample server CRM server.
  title: Golang CRM Swagger Documentation
  version: "1.0"
paths:
...
  /api/banner:
    get:
      consumes:
      - application/json
      description: this api is get banner
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.Banner'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      summary: Get banner
      tags:
      - Banner
    post:
      consumes:
      - application/json
      description: this api is create new banner
      parameters:
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: position
        type: integer
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Create new banner
      tags:
      - Banner
  /api/banner/{id}:
    delete:
      consumes:
      - application/json
      description: this api is to delete banner
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: DELETE banner
      tags:
      - Banner
    get:
      consumes:
      - application/json
      description: this api is Update banner
      parameters:
      - description: banner id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Update banner
      tags:
      - Banner
    put:
      consumes:
      - application/json
      description: this api is Update banner
      parameters:
      - description: category id
        in: path
        name: id
        required: true
        type: string
      - in: formData
        name: descriptionEn
        type: string
      - in: formData
        name: descriptionRu
        type: string
      - in: formData
        name: descriptionUz
        type: string
      - in: formData
        name: nameEn
        type: string
      - in: formData
        name: nameRu
        type: string
      - in: formData
        name: nameUz
        type: string
      - in: formData
        name: position
        type: integer
      - description: file
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/controller.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.response'
      security:
      - BearerAuth: []
      summary: Update banner
      tags:
      - Banner
...
securityDefinitions:
  BearerAuth:
    description: AUTH.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"